+++ spec/operators/observeOn-spec.ts
@@ -104,21 +104,21 @@ describe('Observable.prototype.observeOn', () => {
       .observeOn(Rx.Scheduler.asap)
       .subscribe(
         x => {
-          const observeOnSubscriber = subscription._subscriptions[0]._innerSub;
+          const observeOnSubscriber = subscription._subscriptions[0];
           expect(observeOnSubscriber._subscriptions.length).to.equal(2); // 1 for the consumer, and one for the notification
-          expect(observeOnSubscriber._subscriptions[1]._innerSub.state.notification.kind)
+          expect(observeOnSubscriber._subscriptions[1].state.notification.kind)
             .to.equal('N');
-          expect(observeOnSubscriber._subscriptions[1]._innerSub.state.notification.value)
+          expect(observeOnSubscriber._subscriptions[1].state.notification.value)
             .to.equal(x);
           results.push(x);
         },
         err => done(err),
         () => {
           // now that the last nexted value is done, there should only be a complete notification scheduled
-          const observeOnSubscriber = subscription._subscriptions[0]._innerSub;
+          const observeOnSubscriber = subscription._subscriptions[0];
           expect(observeOnSubscriber._subscriptions.length).to.equal(2); // 1 for the consumer, one for the complete notification
           // only this completion notification should remain.
-          expect(observeOnSubscriber._subscriptions[1]._innerSub.state.notification.kind)
+          expect(observeOnSubscriber._subscriptions[1].state.notification.kind)
             .to.equal('C');
           // After completion, the entire _subscriptions list is nulled out anyhow, so we can't test much further than this.
           expect(results).to.deep.equal([1, 2, 3]);