+++ src/Subscription.ts
@@ -40,7 +40,9 @@ export class Subscription implements ISubscription {
    */
   public closed: boolean = false;
 
-  private _subscriptions: ISubscription[];
+  protected _parent: Subscription = null;
+  protected _parents: Subscription[] = null;
+  private _subscriptions: ISubscription[] = null;
 
   /**
    * @param {function(): void} [unsubscribe] A function describing how to
@@ -66,11 +68,26 @@ export class Subscription implements ISubscription {
       return;
     }
 
+    let { _parent, _parents, _unsubscribe, _subscriptions } = (<any> this);
+
     this.closed = true;
+    this._parent = null;
+    this._parents = null;
+    // null out _subscriptions first so any child subscriptions that attempt
+    // to remove themselves from this subscription will noop
+    this._subscriptions = null;
 
-    const { _unsubscribe, _subscriptions } = (<any> this);
+    let index = -1;
+    let len = _parents ? _parents.length : 0;
 
-    (<any> this)._subscriptions = null;
+    // if this._parent is null, then so is this._parents, and we
+    // don't have to remove ourselves from any parent subscriptions.
+    while (_parent) {
+      _parent.remove(this);
+      // if this._parents is null or index >= len,
+      // then _parent is set to null, and the loop exits
+      _parent = ++index < len && _parents[index] || null;
+    }
 
     if (isFunction(_unsubscribe)) {
       let trial = tryCatch(_unsubscribe).call(this);
@@ -85,8 +102,8 @@ export class Subscription implements ISubscription {
 
     if (isArray(_subscriptions)) {
 
-      let index = -1;
-      const len = _subscriptions.length;
+      index = -1;
+      len = _subscriptions.length;
 
       while (++index < len) {
         const sub = _subscriptions[index];
@@ -138,27 +155,33 @@ export class Subscription implements ISubscription {
       return this;
     }
 
-    let sub = (<Subscription> teardown);
+    let subscription = (<Subscription> teardown);
 
     switch (typeof teardown) {
       case 'function':
-        sub = new Subscription(<(() => void) > teardown);
+        subscription = new Subscription(<(() => void) > teardown);
       case 'object':
-        if (sub.closed || typeof sub.unsubscribe !== 'function') {
-          return sub;
+        if (subscription.closed || typeof subscription.unsubscribe !== 'function') {
+          return subscription;
         } else if (this.closed) {
-          sub.unsubscribe();
-          return sub;
+          subscription.unsubscribe();
+          return subscription;
+        } else if (typeof subscription._addParent !== 'function' /* quack quack */) {
+          const tmp = subscription;
+          subscription = new Subscription();
+          subscription._subscriptions = [tmp];
         }
         break;
       default:
         throw new Error('unrecognized teardown ' + teardown + ' added to Subscription.');
     }
 
-    const childSub = new ChildSubscription(sub, this);
-    this._subscriptions = this._subscriptions || [];
-    this._subscriptions.push(childSub);
-    return childSub;
+    const subscriptions = this._subscriptions || (this._subscriptions = []);
+
+    subscriptions.push(subscription);
+    subscription._addParent(this);
+
+    return subscription;
   }
 
   /**
@@ -168,16 +191,7 @@ export class Subscription implements ISubscription {
    * @return {void}
    */
   remove(subscription: Subscription): void {
-
-    // HACK: This might be redundant because of the logic in `add()`
-    if (subscription == null   || (
-        subscription === this) || (
-        subscription === Subscription.EMPTY)) {
-      return;
-    }
-
-    const subscriptions = (<any> this)._subscriptions;
-
+    const subscriptions = this._subscriptions;
     if (subscriptions) {
       const subscriptionIndex = subscriptions.indexOf(subscription);
       if (subscriptionIndex !== -1) {
@@ -185,17 +199,21 @@ export class Subscription implements ISubscription {
       }
     }
   }
-}
 
-export class ChildSubscription extends Subscription {
-  constructor(private _innerSub: ISubscription, private _parent: Subscription) {
-    super();
-  }
-
-  _unsubscribe() {
-    const { _innerSub, _parent } = this;
-    _parent.remove(this);
-    _innerSub.unsubscribe();
+  private _addParent(parent: Subscription) {
+    let { _parent, _parents } = this;
+    if (!_parent || _parent === parent) {
+      // If we don't have a parent, or the new parent is the same as the
+      // current parent, then set this._parent to the new parent.
+      this._parent = parent;
+    } else if (!_parents) {
+      // If there's already one parent, but not multiple, allocate an Array to
+      // store the rest of the parent Subscriptions.
+      this._parents = [parent];
+    } else if (_parents.indexOf(parent) === -1) {
+      // Only add the new parent to the _parents list if it's not already there.
+      _parents.push(parent);
+    }
   }
 }